requirements.txt
streamlit
pandas
numpy
yfinance
scikit-learn
plotly
requests
alpaca-trade-api
python-dotenv
schedule
